Made changes in the collision detection logic

+++ headers/vehicles.h
@@ -4,7 +4,7 @@
 
 
 
-void drawTank(double x,double y,double z,double delta_x,double delta_y,double delta_z, double turret_Front_Elevation, double turret_Side_Elevation, double FireBallRad, double turnAngle);
+void drawTank(double x,double y,double z,double delta_x,double delta_y,double delta_z, double turret_Front_Elevation, double turret_Side_Elevation, double FireBallRad, int turnAngle);
 void Plane(double x, double y, double z, double delta_x, double delta_y, double delta_z, double horizontal_tilt, double vertical_tilt, double lateral, double blast_rad);
